requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
PyJWT==2.3.0
pytz==2022.1
requests==2.27.1
twilio==7.8.1
urllib3==1.26.9
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
twilio==7.8.1
urllib3==1.26.9
